+++ core/src/test/java/com/hortonworks/iotas/storage/AbstractStoreManagerTest.java
@@ -4,7 +4,7 @@ import com.google.common.collect.Lists;
 import com.hortonworks.iotas.catalog.DataFeed;
 import com.hortonworks.iotas.catalog.DataSource;
 import com.hortonworks.iotas.catalog.Device;
-import com.hortonworks.iotas.catalog.File;
+import com.hortonworks.iotas.catalog.FileInfo;
 import com.hortonworks.iotas.catalog.ParserInfo;
 import com.hortonworks.iotas.catalog.StreamInfo;
 import com.hortonworks.iotas.catalog.Tag;
@@ -315,8 +315,8 @@ public abstract class AbstractStoreManagerTest {
             }};
         }
 
-        protected File createFiles(Long id, String name) {
-            File file = new File();
+        protected FileInfo createFiles(Long id, String name) {
+            FileInfo file = new FileInfo();
             file.setId(id);
             file.setName(name);
             file.setStoredFileName("/tmp/parser.jar");